Source code was synced with Orthodox game framework 3.7.6

+++ SDGF.h
@@ -400,7 +400,7 @@ class Sprite:public Canvas
 class Tileset:public Surface
 {
  private:
- unsigned long int offset;
+ size_t offset;
  unsigned long int tile_width;
  unsigned long int tile_height;
  unsigned long int rows;
